backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
tweepy==4.14.0
python-dotenv==1.0.0
pydantic==2.5.0
sqlalchemy==2.0.23
aiosqlite==0.19.0
apscheduler==3.10.4
python-multipart==0.0.6
httpx==0.25.2
pandas==2.1.4
matplotlib==3.8.2
plotly==5.18.0
